Null check validation done and Readme.md updated

+++ Assignment 8/Controllers/userControllers.js
@@ -2,6 +2,7 @@ const e = require("express");
 const asyncHandler = require("express-async-handler");
 const { findByIdAndDelete, findOneAndDelete } = require("../models/userModel");
 const User = require("../models/userModel");
+const bcrypt = require("bcryptjs");
 
 const editUser = asyncHandler(async (req, res) => {
   //   console.log("something happened");
@@ -12,38 +13,55 @@ const editUser = asyncHandler(async (req, res) => {
   let regexPassword = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{8,16}$/;
   let regexName = /^(?=.{3,20}$)(?![_.])(?!.*[_.]{2})[a-zA-Z ]+(?<![_.])$/;
   let regexEmail = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
+  let regexEmpty = /^$/;
+  let notEmpty = false;
+  if (
+    upname.trim().match(regexEmpty) ||
+    getEmail.trim().match(regexEmpty) ||
+    uppassword.trim().match(regexEmpty)
+  ) {
+    res.status(400).send({
+      message: "Please make sure to enter valid Full Name, Email and Password",
+    });
+  } else {
+    notEmpty = true;
+  }
 
-  if (!upname.trim().match(regexName)) {
-    res.status(400).send({
-      message:
-        "New Full Name should have atleast 3 characters without any special characters",
-    });
-  }
-  if (!getEmail.trim().match(regexEmail)) {
-    res.status(400).send({
-      message: "Email address is not valid",
-    });
-  }
-  if (!uppassword.trim().match(regexPassword)) {
-    res.status(400).send({
-      message:
-        "New Passwords are 8-16 characters with uppercase letters, lowercase letters and at least one number",
-    });
-  }
-  const userUpdate = await User.findOneAndUpdate(
-    { email: getEmail },
-    { $set: { name: upname, password: uppassword } }
-  )
-    .then((data) => {
-      if (data == null) {
+  if (notEmpty) {
+    if (!upname.trim().match(regexName)) {
+      res.status(400).send({
+        message:
+          "New Full Name should have atleast 3 characters without any special characters",
+      });
+    }
+    if (!getEmail.trim().match(regexEmail)) {
+      res.status(400).send({
+        message: "Email address is not valid",
+      });
+    }
+    if (!uppassword.trim().match(regexPassword)) {
+      res.status(400).send({
+        message:
+          "New Passwords are 8-16 characters with uppercase letters, lowercase letters and at least one number",
+      });
+    }
+    const salt = await bcrypt.genSalt(10);
+    uppassword = await bcrypt.hash(uppassword, salt);
+    const userUpdate = await User.findOneAndUpdate(
+      { email: getEmail },
+      { $set: { name: upname, password: uppassword } }
+    )
+      .then((data) => {
+        if (data == null) {
+          res.status(400).send({ message: "User does not exist" });
+        } else {
+          res.status(201).send({ message: "User Updated" });
+        }
+      })
+      .catch((err) => {
         res.status(400).send({ message: "User does not exist" });
-      } else {
-        res.status(201).send({ message: "User Updated" });
-      }
-    })
-    .catch((err) => {
-      res.status(400).send({ message: "User does not exist" });
-    });
+      });
+  }
   //   console.log(upid, upname, uppassword);
 });
 
@@ -53,45 +71,60 @@ const registerUser = asyncHandler(async (req, res) => {
   let regexPassword = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{8,16}$/;
   let regexName = /^(?=.{3,20}$)(?![_.])(?!.*[_.]{2})[a-zA-Z ]+(?<![_.])$/;
   let regexEmail = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
-
-  if (!name.trim().match(regexName)) {
+  let regexEmpty = /^$/;
+  let notEmpty = false;
+  if (
+    name.trim().match(regexEmpty) ||
+    email.trim().match(regexEmpty) ||
+    password.trim().match(regexEmpty)
+  ) {
     res.status(400).send({
-      message:
-        "Full Name should have atleast 3 characters without any special characters",
-    });
-  } else if (!email.trim().match(regexEmail)) {
-    res.status(400).send({
-      message: "Email address is not valid",
-    });
-  } else if (!password.trim().match(regexPassword)) {
-    res.status(400).send({
-      message:
-        "Passwords are 8-16 characters with uppercase letters, lowercase letters and at least one number",
+      message: "Please make sure to enter valid Full Name, Email and Password",
     });
   } else {
-    const userExists = await User.findOne({ email });
+    notEmpty = true;
+  }
 
-    if (userExists) {
+  if (notEmpty) {
+    if (!name.trim().match(regexName)) {
       res.status(400).send({
-        message: "User already exists",
+        message:
+          "Full Name should have atleast 3 characters without any special characters",
       });
-    }
-
-    const user = await User.create({
-      name,
-      email,
-      password,
-    });
-
-    if (user) {
-      res.status(201).json({
-        _id: user._id,
-        name: user.name,
-        email: user.email,
+    } else if (!email.trim().match(regexEmail)) {
+      res.status(400).send({
+        message: "Email address is not valid",
+      });
+    } else if (!password.trim().match(regexPassword)) {
+      res.status(400).send({
+        message:
+          "Passwords are 8-16 characters with uppercase letters, lowercase letters and at least one number",
       });
     } else {
-      res.status(400);
-      throw new Error("Error Occured!!");
+      const userExists = await User.findOne({ email });
+
+      if (userExists) {
+        res.status(400).send({
+          message: "User already exists",
+        });
+      }
+
+      const user = await User.create({
+        name,
+        email,
+        password,
+      });
+
+      if (user) {
+        res.status(201).json({
+          _id: user._id,
+          name: user.name,
+          email: user.email,
+        });
+      } else {
+        res.status(400);
+        throw new Error("Error Occured!!");
+      }
     }
   }
 });
@@ -102,23 +135,34 @@ const deleteUser = asyncHandler(async (req, res) => {
   // console.log(param_delemail, delemail);
 
   let regexEmail = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
-
-  if (!delemail.trim().match(regexEmail)) {
+  let regexEmpty = /^$/;
+  let notEmpty = false;
+  if (delemail.trim().match(regexEmpty)) {
     res.status(400).send({
-      message: "Email address is not valid",
+      message: "Please make sure to enter valid Email",
     });
   } else {
-    const deleteUser = await User.findOneAndDelete({ email: delemail })
-      .then((data) => {
-        if (data == null) {
-          res.status(400).send({ message: "User does not exist" });
-        } else {
-          res.status(201).send({ message: "User Deleted" });
-        }
-      })
-      .catch((err) => {
-        res.status(400).send({ message: "Something went wrong" });
+    notEmpty = true;
+  }
+
+  if (notEmpty) {
+    if (!delemail.trim().match(regexEmail)) {
+      res.status(400).send({
+        message: "Email address is not valid",
       });
+    } else {
+      const deleteUser = await User.findOneAndDelete({ email: delemail })
+        .then((data) => {
+          if (data == null) {
+            res.status(400).send({ message: "User does not exist" });
+          } else {
+            res.status(201).send({ message: "User Deleted" });
+          }
+        })
+        .catch((err) => {
+          res.status(400).send({ message: "Something went wrong" });
+        });
+    }
   }
 });
